worker/requirements.txt
flask==3.0.2
redis==5.0.1
pymongo==4.6.1
python-dotenv==1.0.1
rq==1.16.1
